requirements.txt
fastapi
uvicorn
requests
python-dotenv
msal
azure-storage-blob
pyjwt
# fastapi
# uvicorn
# gunicorn
# requests
# msal
# python-dotenv
# azure-storage-blob
# pandas
# openpyxl

